requirements.txt
Flask==3.1.3
Flask-WTF==1.3.0
SQLAlchemy==2.0.48
alembic==1.18.4
psycopg[binary]==3.3.3
psycopg2-binary==2.9.12
redis==7.4.0
rq==2.8.0
requests==2.33.0
pandas==3.0.1
gunicorn==25.1.0
Flask-Login==0.6.3
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-flask==1.3.0
Flask-APScheduler==1.13.1
beautifulsoup4==4.14.3
curl_cffi==0.15.0
scrapling==0.4.1
playwright==1.58.0
browserforge==1.2.4
patchright==1.58.1
msgspec==0.20.0
nh3==0.3.5
argostranslate==1.11.0
openai==2.32.0
Pillow==12.2.0
rembg==2.0.75
